RUMD: when checkpoint file is found, we do not restart from final.xyz
Note: we should make sure checkpoint is written at the end (run_end?).
The previous behavior made it impossible to prolong a simulation, because
final.xyz existed and therefore the simulation stopped even though the new
condition was not met. This makes it clear that run_pre is not responsible
for stopping the simulation as a general rule.