Make trajectory.steps a property
Trajectory.__init__() need not set steps anymore. This is done via the
read_steps() method, which replaces the various _setup_steps() private
methods that were already there in the xyz hierarchy. This is a slight
optimization.

In principle, this is backward compatible: setting steps in __init__
still works and it circumvents read_steps().