Add callbacks to Optimization and W-tolerance stop condition
The tolerances for Optimization and for the backend are different. The
former tracks the force norm square (W) while the latter are
backend-dependent.

Optimization now inherits from Simulation, since the bulk of the
behavior is the same.

Optimization backends must set a reached_steps attribute at the end of
their run() method. This allows to distinguish the case where the
optimization ended because it reached the maximum number of steps or
the backend tolerance.