Fix behavior of TrajectoryRam with inplace modifications of positions/velocities
We deepcopy the particles in TrajectoryRam to prevent inplace
modifications to numpy arrays like positions and velocities to
propagate to the systems stored in ram.

This will incur in some performance loos by it is safer. As an
optimization one may add a switch between shallow and deep copy
somewhere.

This may fail if the System object is too complex does not support
deepcopy.