Make step parameter optional in trajectory write(); do not alloww overwriting frames by default.
By default, we do not allow anymore overwiting a frame, unless the
_overwrite trajectory attribute is True. This avoids checking if a
step is already there. Also this makes sense because standard
trajectory formats do not allow overwriting.

TrajectoryRam and TrajectoryRamFull have overwrite mode on by default.

As a result, writing trajectory samples is safer and little a bit
faster.