Add view mode (mode='v') to TrajectoryRamFull
This switch allows to specify the behavior of TrajectoryRamFull. By
default it returns a view on the System in memory, which allows
in-place modifications. The 'r' mode instead returns a deep copy of
the system. This is the default for trajectories on disk.

TrajectoryRam should soon be replaced by TrajectoryRamFull.