Fix rare numpy formatting in xyz files
Closing xyz files resets the numpy string function. If this occurs
between two calls to write_sample(), this could screw up
formatting. To prevent this, we set the string function at every
call. This may incur in a slight overhead, but at least it is safe as
long as we stay single thread.