Fix simulation callback behavior when adding the same callback with different arguments
This was the occasion for a large refactoring of the callbacks
handling in Simulation. The observer is now more clearly defined as a
tuple of (id, callback, args, kwargs). The missing bit was the id for
the observers, which allows one to remove individual observers from
the list.

The commit should be 100% backward compatible unless client classes
mess around with underscored variables.